requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.9.11
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.24.1
h5py==2.10.0
joblib==0.14.0
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rectified-adam==0.17.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
opt-einsum==3.1.0
pandas==0.25.1
patsy==0.5.1
Pillow==6.2.0
protobuf==3.11.2
pydicom==1.3.0
pydot==1.4.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.2
seaborn==0.9.0
six==1.13.0
statsmodels==0.11.1
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
text-unidecode==1.3
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
